+++ NEWS.md
@@ -1,5 +1,6 @@
 # fluxnet (development version)
 
+* Added `flux_citations()` to generate site-level citations either as plain text or BibTex entries.
 * Fixed a bug where hourly data wasn't being extracted or read in along with half-hourly data ([#73](https://github.com/EcosystemEcologyLab/fluxnet-package/issues/73))
 
 # fluxnet 0.3.2